alteracao na deleção das vagas

+++ src/Controller/DeleteJobPostingController.java
@@ -9,24 +9,38 @@ public class DeleteJobPostingController {
 
     private final Views views;
     private JobPostingRepository repository;
+    private final ApplicationRepository applicationRepository;
 
-    public DeleteJobPostingController(Views views, JobPostingRepository repository) {
+    public DeleteJobPostingController(Views views, JobPostingRepository repository, ApplicationRepository applicationRepository) {
         this.views = views;
         this.repository = repository;
+        this.applicationRepository = applicationRepository;
     }
 
     public void delete()
     {
         int id = views.deleteJobPostingView.show();
+
         JobPosting jobPosting = repository.findbyId(id);
+
         if(jobPosting==null)
         {
             views.deleteJobPostingView.showFail();
-        }else
+            return;
+        }
+        if (applicationRepository.hasApplications(jobPosting))
         {
-            repository.delete(jobPosting);
-            views.deleteJobPostingView.showSuccess();
+            views.deleteJobPostingView.ShowJobPostingHasApplications();
+            return;
         }
 
+        if (!views.deleteJobPostingView.confirmDelete())
+        {
+            return;
+        }
+
+            repository.delete(jobPosting);
+
+            views.deleteJobPostingView.showSuccess();
     }
 }
